backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
numpy
joblib
scikit-learn
xgboost
pydantic
